restructured folders for clarity. index.js has to stay inside src or npm start can't find it. Reorganized css so it makes more sense.
