+++ src/css/ImageUpload.css
@@ -1,4 +1,4 @@
-.fileInput{
+.fileInput {
   border-bottom: 2px
   solid gray;
   border-right: 2px solid grey;
@@ -8,7 +8,8 @@
   margin: 10px;
   cursor: pointer;
 }
-.imgPreview{
+
+.imgPreview {
   position: absolute;
   left: 250px;
   padding: 5px;
@@ -20,17 +21,19 @@
   border-top: 2px solid lightgray;
   border-bottom: 2px solid lightgray;
 }
-.uploadedImg {
-    width: 100%;
-    height: 100%;
-  }
 
-.previewText{
+.uploadedImg {
+  width: 100%;
+  height: 100%;
+}
+
+.previewText {
   width: 100%;
   margin-top: 10px;
   text-align: center;
 }
-.submitButton{
+
+.submitButton {
   padding: 12px;
   background: white;
   border-radius: 10px;
@@ -38,8 +41,9 @@
   font-size: 8pt;
   cursor: pointer;
 }
-  &:hover{
-    background: #efefef
+
+&:hover{
+  background: #efefef
 }
 
 .previewComponent {
