+++ src/css/index.css
@@ -1,0 +1,6 @@
+body {
+  background-color: Snow;
+  font-family: sans-serif;
+  margin: 0;
+  padding: 0;
+}
