--- src/index.css
@@ -1,18 +1,0 @@
-body {
-  margin: 0;
-  padding: 0;
-  font-family: sans-serif;
-  background-color:Snow;
-  text-align: center;
-  font-family: sans-serif;
-
-}
-#headerid{
-  background: SteelBlue;
-
-}
-
-h2{
-  margin-left: 15px;
-  text-align: center;
-}
