+++ src/js/ImageUpload.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import './ImageUpload.css'
+import '../css/ImageUpload.css'
 
 class ImageUpload extends Component {
   constructor(props) {
